agents/requirements.txt
paho-mqtt==2.1.0
PyYAML==6.0.2
pydantic==2.10.6
aiohttp>=3.9.0
